Break out into separate files, bundle with browserify
