+++ lib/polyfill-is-array.js
@@ -1,0 +1,6 @@
+'use strict';
+
+module.exports = Array.isArray || function isArray(o) {
+  // older browsers
+  return Object.prototype.toString.call(o) === '[object Array]';
+};
