+++ package.json
@@ -1,10 +1,27 @@
 {
   "name": "async-json",
-  "version": "0.0.2",
+  "version": "1.0.0",
   "description": "An asynchronous version of JSON.stringify",
-  "keywords": ["json", "stringify", "async"],
+  "keywords": [
+    "json",
+    "stringify",
+    "async"
+  ],
+  "main": "index",
   "repository": "git://github.com/ckknight/async-json.git",
   "author": "Cameron Kenneth Knight <ckknight@gmail.com> (http://ckknight.com)",
-  "dependencies": {},
-  "main": "index"
+  "dependencies": {
+  },
+  "devDependencies": {
+    "assert": "^1.3.0",
+    "bluebird": "^2.10.0",
+    "browserify": "^11.1.0",
+    "expresso": "^0.9.2"
+  },
+  "scripts": {
+    "browserify": "./node_modules/.bin/browserify ./lib/async-json.js -o ./dist/async-json.js",
+    "prepublish": "npm run browserify",
+    "unit": "node_modules/expresso/bin/expresso ./test/*.test.js",
+    "test": "npm run unit"
+  }
 }
